offer_app/requirements-test.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.1
dnspython==2.5.0
email-validator==2.1.0.post1
Faker==22.6.0
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.4
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pytest==8.0.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
sqlmodel==0.0.14
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0
watchfiles==0.21.0
websockets==12.0
